refactor: renamed models
