+++ src/models/apiConfiguration.model.ts
