+++ src/models/application.model.ts
@@ -2,9 +2,9 @@ import { AccessibleRecordModel, accessibleRecordsPlugin } from '@casl/mongoose';
 import mongoose, { Schema, Document } from 'mongoose';
 import { status } from '../const/enumTypes';
 import { addOnBeforeDeleteMany } from '../utils/models/deletion';
-import { Page } from './page';
-import { Role } from './role';
-import { Channel } from './channel';
+import { Page } from './page.model';
+import { Role } from './role.model';
+import { Channel } from './channel.model';
 
 /** Application documents interface declaration */
 export interface Application extends Document {
