+++ src/models/form.model.ts
@@ -2,10 +2,10 @@ import { AccessibleRecordModel, accessibleRecordsPlugin } from '@casl/mongoose';
 import mongoose, { Schema, Document } from 'mongoose';
 import { addOnBeforeDeleteMany } from '../utils/models/deletion';
 import { status } from '../const/enumTypes';
-import { Channel } from './channel';
-import { layoutSchema } from './layout';
-import { Version } from './version';
-import { Record } from './record';
+import { Channel } from './channel.model';
+import { layoutSchema } from './layout.model';
+import { Version } from './version.model';
+import { Record } from './record.model';
 import { getGraphQLTypeName } from '../utils/validators';
 
 /** Form documents interface declaration */
