+++ src/models/index.ts
@@ -1,24 +1,24 @@
-export * from './apiConfiguration';
-export * from './application';
-export * from './channel';
-export * from './client';
-export * from './dashboard';
-export * from './form';
-export * from './layout';
-export * from './notification';
-export * from './page';
-export * from './permission';
-export * from './positionAttribute';
-export * from './positionAttributeCategory';
-export * from './pullJob';
-export * from './record';
-export * from './referenceData';
-export * from './resource';
-export * from './role';
-export * from './setting';
-export * from './step';
-export * from './user';
-export * from './version';
-export * from './workflow';
-export * from './history';
-export * from './group';
+export * from './apiConfiguration.model';
+export * from './application.model';
+export * from './channel.model';
+export * from './client.model';
+export * from './dashboard.model';
+export * from './form.model';
+export * from './layout.model';
+export * from './notification.model';
+export * from './page.model';
+export * from './permission.model';
+export * from './positionAttribute.model';
+export * from './positionAttributeCategory.model';
+export * from './pullJob.model';
+export * from './record.model';
+export * from './referenceData.model';
+export * from './resource.model';
+export * from './role.model';
+export * from './setting.model';
+export * from './step.model';
+export * from './user.model';
+export * from './version.model';
+export * from './workflow.model';
+export * from './history.model';
+export * from './group.model';
