+++ src/models/user.model.ts
@@ -1,7 +1,7 @@
 import { AccessibleRecordModel, accessibleRecordsPlugin } from '@casl/mongoose';
 import mongoose, { Schema, Document } from 'mongoose';
 import { AppAbility } from '../security/defineUserAbility';
-import { PositionAttribute } from './positionAttribute';
+import { PositionAttribute } from './positionAttribute.model';
 
 /** Mongoose user schema definition */
 const userSchema = new Schema({
