+++ src/utils/history/recordHistory.ts
@@ -1,5 +1,8 @@
 import { Record, User, Role, ReferenceData } from '../../models';
-import { Change, RecordHistory as RecordHistoryType } from 'models/history';
+import {
+  Change,
+  RecordHistory as RecordHistoryType,
+} from '../../models/history.model';
 import { AppAbility } from 'security/defineUserAbility';
 import dataSources, { CustomAPI } from '../../server/apollo/dataSources';
 import { isArray, memoize } from 'lodash';
