+++ src/utils/schema/resolvers/Meta/getMetaUsersResolver.ts
@@ -1,4 +1,4 @@
-import { User } from '../../../../models/user';
+import { User } from '../../../../models';
 import mongoose from 'mongoose';
 
 /**
